+++ doc/changes.md
@@ -19,6 +19,8 @@ development source code and as such may not be routinely kept up to date.
 (v-next-improvements)=
 ### Improvements
 
+* `nextstrain setup --list` now lists the available pathogens for setup.
+  ([#502](https://github.com/nextstrain/cli/issues/502)
 * Added the following to the list of "well-known" environment variables that are
   automatically passed thru to runtimes: